Restructure modules into mocksafe.core and mocksafe.apis.bdd
