+++ mocksafe/__init__.py
@@ -12,13 +12,23 @@ See:
  - :meth:`mocksafe.that`
  - :meth:`mocksafe.spy`
 """
-from mocksafe.mock import mock, mock_module, mock_reset
-from mocksafe.when_then import when, stub
-from mocksafe.that import that, spy
-
-from mocksafe.when_then import WhenStubber, MatchCallStubber, LastCallStubber
-from mocksafe.mock_property import MockProperty, PropertyStubber
-from mocksafe.that import MockCalls
+from mocksafe.core import (
+    MockProperty,
+    mock,
+    mock_module,
+    mock_reset,
+)
+from mocksafe.apis.bdd import (
+    WhenStubber,
+    MatchCallStubber,
+    LastCallStubber,
+    PropertyStubber,
+    MockCalls,
+    when,
+    stub,
+    that,
+    spy,
+)
 
 
 __version__ = "0.6.0-alpha"
