+++ mocksafe/core/__init__.py
@@ -1,0 +1,2 @@
+from mocksafe.core.mock import mock, mock_module, mock_reset
+from mocksafe.core.mock_property import MockProperty
